[WEB-822] Rename therapySettingsFormSteps to therapySettingsFormStep
